jquery locally, reset password button in admin console, secured the post signup and updated the registration_enabled

+++ project/auth.py
@@ -1,9 +1,10 @@
 from flask_login import login_user, login_required, logout_user, current_user
 from flask import Blueprint, render_template, redirect, url_for, request, flash
 from werkzeug.security import generate_password_hash, check_password_hash
-from .models import User
+from .models import User, RegisterLink
 from . import db
 from .models import Config
+
 
 auth = Blueprint('auth', __name__)
 
@@ -32,8 +33,8 @@ def login_post():
     return redirect(url_for('main.index'))
 
 
-@auth.route('/signup')
-def signup():
+@auth.route('/signuup')
+def signuup():
     #config = Config.query.first()
     try:
         config = Config.query.order_by(Config.id.desc()).first()
@@ -53,6 +54,24 @@ def signup():
         flash('The user registration is disabled')
         return redirect(url_for('main.index'))
 
+
+@auth.route('/signup')
+def signup():
+    #config = Config.query.first()
+    try:
+        config = Config.query.order_by(Config.id.desc()).first()
+
+        if config.registration_enabled:
+            enabled = True
+        else:
+            enabled = False
+    except:
+        enabled = False
+
+    return render_template('signup.html',registration_enabled=enabled)
+
+
+
 @auth.route('/signup', methods=['POST'])
 def signup_post():
 
@@ -60,21 +79,50 @@ def signup_post():
     email = request.form.get('email')
     name = request.form.get('name')
     password = request.form.get('password')
+    token = request.form.get('token')
 
-    user = User.query.filter_by(email=email).first() # if this returns a user, then the email already exists in database
+    valid=False
 
-    is_admin=False
+    #Check if token is valid and return true or false
+    if token:
+        try:
+            result = RegisterLink.query.filter_by(link_token=token).one()
+            RegisterLink.query.filter_by(link_token=token).delete()
+            db.session.commit()
+            valid=True
+        except:
+            valid=False
+            return "Invalid or expired token"
 
-    if user: # if a user is found, we want to redirect back to signup page so user can try again
-        flash('Email address already exists')
-        return redirect(url_for('auth.signup'))
+    #Check if registration or the token are enabled or valid and return true or false
+    try:
+        config = Config.query.order_by(Config.id.desc()).first()
+        if config.registration_enabled or valid==True:
+            enabled = True
+        else:
+            enabled = False
+    except:
+        enabled = False
 
-    # create a new user with the form data. Hash the password so the plaintext version isn't saved.
-    new_user = User(email=email, name=name, password=generate_password_hash(password, method='sha256'),is_admin=is_admin)
 
-    # add the new user to the database
-    db.session.add(new_user)
-    db.session.commit()
+    #If token or registration are valid, then create the user.
+
+    if enabled == True:
+
+        user = User.query.filter_by(email=email).first() # if this returns a user, then the email already exists in database
+
+        is_admin=False
+
+        if user: # if a user is found, we want to redirect back to signup page so user can try again
+            flash('Email address already exists')
+            return redirect(url_for('auth.signup'))
+
+        # create a new user with the form data. Hash the password so the plaintext version isn't saved.
+        new_user = User(email=email, name=name, password=generate_password_hash(password, method='sha256'),is_admin=is_admin)
+
+        # add the new user to the database
+        db.session.add(new_user)
+        db.session.commit()
 
     return redirect(url_for('auth.login'))
 
@@ -178,21 +226,55 @@ def users_edit():
 @login_required
 def users_delete():
 
-    user_id = request.form.get('deleteUserIdInput')
-    print(user_id)
-    try:
-        User.query.filter_by(id=user_id).delete()
-        db.session.commit()
-        flash("User deleted")
-    except:
-        flash("Error deleting user")
+    if current_user.is_admin:
 
-    return redirect(url_for('main.admin'))
+        user_id = request.form.get('deleteUserIdInput')
 
-def db_setup(email,name,password,is_admin):
+        try:
+            User.query.filter_by(id=user_id).delete()
+            db.session.commit()
+            flash("User deleted")
+        except:
+            flash("Error deleting user")
+
+        return redirect(url_for('main.admin'))
+
+
+@auth.route('/users/reset-password', methods=['POST'])
+@login_required
+def users_reset_password():
+
+    if current_user.is_admin:
+
+        user_id = (request.form.get("userIdInput2"))
+
+        try:
+            password = "ChangeMe"
+            password = generate_password_hash(password, method='sha256')
+
+            user = User.query.filter_by(id=user_id).first()
+            user.password = password
+
+            db.session.commit()
+
+            flash("Password restored")
+        except:
+            flash("Error restoring the password")
+
     
-    admin_user = User(email=email, name=name, password=generate_password_hash(password, method='sha256'),is_admin=is_admin)
+        return redirect(url_for('main.admin'))
 
-    # add the new user to the database
-    db.session.add(admin_user)
-    db.session.commit()
+@auth.route('/sign-up/<token>')
+def register(token):
+    # Retrieve the user from the database using the token
+
+    try:
+
+        result = RegisterLink.query.filter_by(link_token=token).one()
+        print('Valid token')
+
+        return render_template('signup.html', registration_enabled=True, token=token)
+    except:
+        # If the query raises NoResultFound, the ID doesn't exist
+        print('Invalid or expired token')
+        return "Invalid or expired token"
